Created PreProcessing class for preprocessing data
The preprocessing class is used for preprocessing data for the the CNN
and SVM model we are going to use. It extracts mean featrures for the
SVM and extracts signal representational features for the CNN that
can be inputted as a 2D images. It is going to be used to store the
data we have preprocessed, so we do not have to run the preprocessing
steps over and over again.