Add requirements text file
This file is for robots or other machines so they know what libraries
have to be installed in order to run this code. Furthermore, it is for
people that want to know what libraries are required for usage.

The requirements txt will be filed out as external libraries will be
added to the code.